Added a vision...
Heavy modification to the overall look of the game; voxels now present a
LEGO-like bump on the top and dimple on the bottom.
Changed the shader to be far simpler, using a directional light source
for sunlight instead of a point source.
Moved a lot of the matrix calculations out of the main code and into
GLSL; this does result in a lot of redundant calculations per pixel
right now, but offloads it from the CPU.  Might be a way to simplify
this further.