So long, namespaces.
I got rid of couching every single file inside the "gsl::rectopia"
namespace, because (a) it's no longer called Rectopia, and (b) THIS
ISN'T JAVA.  Seriously, there's just no reason for me to namespace this
code, and it was annoying me.

On the other hand, this means I touched EVERY SINGLE FILE.  So there's
that.